Merge logic to interoperate with MDUtils
I added some logic to MDUtils to add my parsing
logic for the double asterisks, as well as fixing
a bug.

The bug would happen if you had an empty line.
The code would try to get the substring of the first
two characters, which throws an exception since there
is no characters at all.